requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
psycopg2-binary>=2.9.9
pgvector>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.12
langchain-core>=0.1.10
langchain-openai>=0.1.0
langgraph>=0.0.20
langchain-ollama>=0.1.0
pypdf>=4.0.0
pdf2image>=1.17.0
Pillow>=10.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.26.0
reportlab>=4.0.0
python-docx>=1.1.0
